Exercise Progress
Yep, the exercise is almost finished already. 

The point of it is to find the binomial coefficients of user-determined variables "n" and "k", in three different ways. One using loops, one using recursion, and one way using recursion with multi-dimensional arrays.

By all means the loop method is the easiest for me to wrap my head around, so that's the only thing I haven't done.

I'll be commenting on and improving some loose ends later.